feat(components): refactor Entrypoint methods
Refactor the `Entrypoint` component to remove the `dataPropertyKey` parameter from the `runControllerMethodWithoutAbortController` and `runControllerMethodWithAbortController` methods. These methods now directly access the data property from the `context` object. Also, update the `runControllerMethod` method to pass the `context` object and optional `abortController` to either `runControllerMethodWithoutAbortController` or `runControllerMethodWithAbortController` depending on the presence of `abortController`.