feat: Add EntrypointBuilder function to create entrypoint options
The EntrypointBuilder function is added to the Entrypoint module to allow creating entrypoint options with the provided CLI, HTTP, and Service entrypoints. This function takes an options object with the following properties:

- cli: A CLIEntrypoint or an array of CLIEntrypoint functions.
- http: An HTTPEntrypoint or an array of HTTPEntrypoint functions.
- service: A ServiceEntrypoint or an array of ServiceEntrypoint functions.

The function returns the options object with the provided entrypoints. This allows creating entrypoints with different configurations in a more readable and organized way.

The commit also includes changes in the App.spec.ts file to use the EntrypointBuilder function instead of directly providing the entrypoints.