feat(core): Add environment variable support to Application class
This commit adds a new method to the `Application` class called `env` which allows users to set environment variables. The method takes in a record of key-value pairs representing the environment variables to set.

Additionally, this commit also adds a new method to the `Provider` class called `getEnv` which allows users to retrieve environment variables. The method takes in the name of the environment variable and an optional default value to return if the environment variable is undefined.

The changes in this commit enable better control and management of environment variables within the application.