feat(Application): add alias path support
Add support for registering path aliases in the Application class. Path aliases can be used to provide shorter and more convenient paths to files or directories in the application. Path aliases can be registered using the `alias` method, which takes an object of alias names and their corresponding paths as input. The aliases can then be used in the application code by prefixing the alias name with `@`.

Example usage:
```typescript
Application.alias({
  '@test': '/path/to/test',
  '@runtime': process.cwd()
});
```

In this example, the `@test` alias is registered with the path `/path/to/test`, and the `@runtime` alias is registered with the current working directory path. These aliases can then be used in the application code like this:

```typescript
import { SomeModule } from '@test/someModule';
import { SomeFile } from '@runtime/someFile';
```

The `@test` alias will be resolved