feat: Add ImageImportOptions class to handle import options
This commit adds a new file `ImageImportOptions.ts` to the `src/components/docker/options` directory. The `ImageImportOptions` class extends the `DTO` class from the `lib/core/DTO.js` module. This class is used to handle import options for importing Docker images from .tar files.

In addition, the `Docker` class in `Docker.ts` is updated to include a new method `importImage` which accepts an `ImageImportOptions` parameter. This method is annotated with `@Accept(ImageImportOptions.required())` to enforce the required presence of `ImageImportOptions` when calling the `importImage` method.

The implementation of the `importImage` method is not yet complete and throws a `not implemented` error.