feat: Add ContainerVolume type
Adds a new type `ContainerVolume` to the `src/components/docker/types/ContainerVolume.ts` file. The `ContainerVolume` type consists of two properties: `hostPath` and `containerPath`, both of type `string`. This type will be used in the `DockerContainer` class to represent the volumes of a Docker container.

This commit addresses the following tasks:
- Create `ContainerVolume` type
- Define `hostPath` property in `ContainerVolume`
- Define `containerPath` property in `ContainerVolume`