feat(docker): Add support for setting container capabilities
Add support for setting container capabilities when creating a Docker container. The `ContainerSettingOptions` now includes a `capabilities` field that accepts an array of `ContainerCapability` values. This allows users to specify additional kernel capabilities for the container.

Example usage:

```typescript
const container = await docker.createContainer({
    image: 'my-image',
    name: 'my-container',
    capabilities: [ContainerCapability.AUDIT_CONTROL, ContainerCapability.NET_ADMIN]
});
```

This commit also includes updates to the `ContainerCommitOptions` and `DockerContainer` classes to support committing a container with a repoTag and pausing the container before committing.