feat: Add framework information provider
Added a new provider called "Information" that retrieves and formats information about the framework. This provider is used in the "info" command of the CommandLineController to display framework information.

The provider retrieves the framework name, version, description, license, and installation path from the package.json file. It also searches for the root path of the project where the working directory is located.

The provider provides methods to format and retrieve various information, such as the framework ASCII logo, framework description, version text, installation path, running level (project or global), and license. It also provides methods to retrieve the project level installed framework version and to print the formatted information.

This commit also includes some changes in the CommandLineController and the CLIApp to use the new provider and display the framework information.