feat: Add SHA256, MD5, and SHA1 hash functions
This commit adds three new hash functions to the project: SHA256, MD5, and SHA1. These functions use both the crypto-js and node:crypto libraries to generate the hash values. If the node:crypto library does not have support for a specific hash algorithm, the crypto-js library is used as a fallback.

The new hash functions are implemented in the following files:
- src/lib/helpers/SHA256.ts
- src/lib/helpers/MD5.ts
- src/lib/helpers/SHA1.ts

These functions take a string message as input and return the hash value as a Buffer.

Additionally, the new hash functions are exported in the src/exports/Helpers.ts file, so they can be accessed from other parts of the project.

This commit also includes some test code in the src/tests/App.spec.ts file to demonstrate the usage of the new hash functions.